[ETCM-759] Move miner tests to 'miners' package
